refactor(backend): extract AgentService core into eight controllers
agent-service.ts 2472 -> 1241 lines. The facade keeps the public API and event emission; each extracted controller owns one concern with a *Hooks interface and never imports the facade: ImageGenRuntime, DeltaBuffer, MailboxSync, BootRecovery, AttachmentGateway, ThreadLifecycle, DrainScheduler, TurnLifecycle.
